+++ HISTORY.rst
@@ -100,6 +100,7 @@
 - Add a tags and filters selection to the template widget.
 - Remove runtime.py modules and move instancing to base class.
   Avoids keeping long lived objects in memory.
+- Consolidate app module loading using AppsModuleLoaderMixin.
 
 3.4.9 (2020-05-26)
 ==================